Create users with identity and authentication in dev console
Dev console created only users, you can't login with just user.  Change
it to also create identity and authentication.  The password for all the
users created in the dev console is "password".  Also the users created
in the dev console are set to be not temp (not new users) and all the
contracts are automatically accepted.