Remove ask_for_email page from the registration workflow
The page was originally created for twitter users because twitter
doesn't give us an email address.  We require email address verification
in order for users to be able to reset their own passwords, but that's
not necessary for twitter users.  So this page is removed from the
registration workflow for now.