Store email on social signup form if social network didn't give us one
For example, if I log in using twitter, twitter doesn't return an
email address, I put in an email address on the social signup form, that
email address will be saved and unverified.

If I log in using facebook, facebook returns an email address, that
email address is saved and verified.