Update en.yml
Updated to the below. I also found a few others that weren't in the original list that I fixed. Thanks! 
Your new log in option as been added!
That log in option is already used by someone else. If that someone is you, remove it from your other account and try again.
That way to log in cannot be added because it is associated to an email address that is already in use!
Can’t log in?
Enter the email or username associated with your OpenStax account, and we will help you log in.
We found several accounts with this email. Please log in using your username
Please check ... for an email listing your usernames, and use one to log in.
Please log in again before continuing.