Rename to sf contact id in secu log SF sync
In the scheduled SF sync, when a user is faculty verifed via salesforce, a security log is created. That security log includes the salesforce contact id in the event data. The problem is that it calls it `salesforce_id` which is inaccurate; it should say `salesforce_contact_id` because that's what it is. Just improves the security logs stream.