Send all user info to SF on lead create
Previously, when a salesforce lead for a user was created, there was no need to send ALL of the User record's fields because upon lead creation, we did not yet have all of that information. But that changed when we started upserting Leads, which means that upon Lead creation, we some times already have all of the user's info, so we want to send it all at that point.