Update Lead source to faculty
When a user signs up as a student but then requests faculty access, and they get access, their salesforce lead is updated. The problem is that their Lead's `source` (which is like their role) remains set to Student. The fix is to update the lead's `source` field to `OSC Faculty` as it does get set when the lead is first created (so just use that constant `CreateSalesforceLead::SALESFORCE_INSTRUCTOR_ROLE`).