CORE-986: Fix color contrast of signup heading (#1645)
* CORE-986: Fix color contrast of signup heading

[CORE-986]
New color contrast ratio is 5.
Also moved the disabled-submit styling to CSS
Also added some JS to control the disabled state of the Continue button on login.

* CORE-1047

Because this is so tiny, I'm rolling it in here.
Fixes contrast of Show/Hide label (which is a problem when the input is in an error state - pink)

* Remove test for blank email; submit is disabled if email is blank